Buffer logging output in a thread
This change is a bit of a lark, it seems better to buffer the logging
output into a single thread and remove the connection and write out
overhead from the responsibility of each worker. The issue here is that
we're trading the write out time for the time it takes to aquire a RW
lock and append your own data to it. Intuitively this should be much
faster for a moderate to small number of workers assuming memory access
is faster than even a local tcp socket.